requirements.txt
flask[async]~=3.1.0
quart
pre-commit
black
python-dotenv==1.0.1
google-cloud-firestore>=2.19.0
pytest~=8.3.4
pytest-asyncio
pydantic~=2.9.2
firebase-admin~=6.6.0


# langchain
langchain
langchain-community~=0.3.14
langchain-core~=0.3.30
langchain-google-community
langchain-google-genai~=2.0.9
google-cloud-discoveryengine
langgraph~=0.2.22

# langfuse
langfuse==2.56.2
protobuf~=4.25.5
injector~=0.22.0